requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
alembic==1.15.2
    # via optuna
annotated-types==0.7.0
    # via pydantic
astroid==3.3.10
    # via pylint
beaupy==3.10.1
black==25.1.0
bleach==6.2.0
    # via panel
bokeh==3.7.3
    # via
    #   holoviews
    #   panel
certifi==2025.4.26
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.2
    # via requests
click==8.2.0
    # via
    #   black
    #   wandb
colorcet==3.1.0
    # via
    #   datashader
    #   holoviews
colorlog==6.9.0
    # via optuna
contourpy==1.3.2
    # via
    #   bokeh
    #   matplotlib
cycler==0.12.1
    # via matplotlib
datashader==0.18.1
dill==0.4.0
    # via pylint
docker-pycreds==0.4.0
    # via wandb
emoji==2.14.1
    # via beaupy
filelock==3.18.0
    # via torch
fonttools==4.58.0
    # via matplotlib
fsspec==2025.3.2
    # via torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via wandb
greenlet==3.2.2
    # via
    #   pynvim
    #   sqlalchemy
holoviews==1.20.2
idna==3.10
    # via requests
imageio==2.37.0
    # via scikit-image
isort==6.0.1
    # via pylint
jinja2==3.1.6
    # via
    #   bokeh
    #   torch
joblib==1.5.0
    # via
    #   pynndescent
    #   scikit-learn
kiwisolver==1.4.8
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
linkify-it-py==2.0.3
    # via panel
llvmlite==0.44.0
    # via
    #   numba
    #   pynndescent
mako==1.3.10
    # via alembic
mambapy==1.2.0
markdown==3.8
    # via panel
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   panel
    #   rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
matplotlib==3.10.3
    # via
    #   scienceplots
    #   seaborn
mccabe==0.7.0
    # via pylint
mdit-py-plugins==0.4.2
    # via panel
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
msgpack==1.1.0
    # via pynvim
multipledispatch==1.0.0
    # via datashader
mypy-extensions==1.1.0
    # via black
narwhals==1.39.0
    # via bokeh
networkx==3.4.2
    # via
    #   scikit-image
    #   torch
numba==0.61.2
    # via
    #   datashader
    #   pynndescent
    #   umap-learn
numpy==2.2.5
    # via
    #   bokeh
    #   contourpy
    #   datashader
    #   holoviews
    #   imageio
    #   matplotlib
    #   numba
    #   optuna
    #   pandas
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   seaborn
    #   tifffile
    #   umap-learn
    #   xarray
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
optuna==4.3.0
packaging==25.0
    # via
    #   black
    #   bokeh
    #   datashader
    #   holoviews
    #   lazy-loader
    #   matplotlib
    #   optuna
    #   panel
    #   scikit-image
    #   xarray
pandas==2.2.3
    # via
    #   bokeh
    #   datashader
    #   holoviews
    #   panel
    #   seaborn
    #   xarray
panel==1.6.3
    # via holoviews
param==2.2.0
    # via
    #   datashader
    #   holoviews
    #   panel
    #   pyct
    #   pyviz-comms
pathspec==0.12.1
    # via black
pillow==11.2.1
    # via
    #   bokeh
    #   imageio
    #   matplotlib
    #   scikit-image
platformdirs==4.3.8
    # via
    #   black
    #   pylint
    #   wandb
polars==1.29.0
protobuf==6.30.2
    # via wandb
psutil==7.0.0
    # via wandb
pyct==0.5.0
    # via datashader
pydantic==2.11.4
    # via wandb
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via rich
pylint==3.3.7
pynndescent==0.5.13
    # via umap-learn
pynvim==0.5.2
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-yakh==0.4.1
    # via
    #   beaupy
    #   questo
pytz==2025.2
    # via pandas
pyviz-comms==3.0.4
    # via
    #   holoviews
    #   panel
pyyaml==6.0.2
    # via
    #   bokeh
    #   optuna
    #   wandb
questo==0.4.1
    # via beaupy
requests==2.32.3
    # via
    #   datashader
    #   panel
    #   wandb
rich==14.0.0
    # via
    #   beaupy
    #   questo
scienceplots==2.1.1
scikit-image==0.25.2
scikit-learn==1.6.1
    # via
    #   pynndescent
    #   umap-learn
scipy==1.15.3
    # via
    #   datashader
    #   pynndescent
    #   scikit-image
    #   scikit-learn
    #   umap-learn
seaborn==0.13.2
sentry-sdk==2.28.0
    # via wandb
setproctitle==1.3.6
    # via wandb
setuptools==80.4.0
    # via
    #   torch
    #   triton
    #   wandb
six==1.17.0
    # via
    #   docker-pycreds
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sqlalchemy==2.0.40
    # via
    #   alembic
    #   optuna
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tifffile==2025.5.10
    # via scikit-image
tomlkit==0.13.2
    # via pylint
toolz==1.0.0
    # via datashader
torch==2.7.0
    # via mambapy
tornado==6.4.2
    # via bokeh
tqdm==4.67.1
    # via
    #   optuna
    #   panel
    #   umap-learn
triton==3.3.0
    # via torch
typing-extensions==4.13.2
    # via
    #   alembic
    #   panel
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspection
    #   wandb
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
uc-micro-py==1.0.3
    # via linkify-it-py
umap-learn==0.5.7
urllib3==2.4.0
    # via
    #   requests
    #   sentry-sdk
wandb==0.19.11
webencodings==0.5.1
    # via bleach
xarray==2025.4.0
    # via datashader
xyzservices==2025.4.0
    # via bokeh
